click.cc: Move argument parsing to a separate file
So it can be reused
